+++ src/Types/CreateQuizType.ts
@@ -1,0 +1,7 @@
+export type CreateQuizType = {
+  title: string;
+  description: string;
+  comment: string;
+  topicId: number;
+  numberOfQuestions: number
+}
