+++ src/api/quizesApi.ts
@@ -1,18 +1,14 @@
 import { FieldValues } from 'react-hook-form';
 import { instance } from './Instance/instance';
 import { CardsType } from '../Types/CardTypes';
+import { CreateQuizType } from "../Types/CreateQuizType";
 
-export type CreateQuestionFieldType = {
-  title: string;
-  description: string;
-  topicId: number;
-};
 export const quizesApi = {
   getQuizes() {
     return instance.get<CardsType[]>(`quiz`);
   },
-  postQuizes(data: FieldValues) {
-    return instance.post<FieldValues>('quiz', data);
+  postQuizes(quizData: CreateQuizType) {
+    return instance.post<CreateQuizType>('quiz', quizData);
   },
   postQuestion(questionData: FieldValues) {
     return instance.post<FieldValues>('questions', questionData);
